chore(weave): extract clickhouse/schema_converters.py from clickhouse_trace_server_batched
Move call, object, and table schema conversion functions into a dedicated
module. Also removes commented-out dead code (_partial_obj_schema_to_ch_obj).
