+++ weave/trace_server/clickhouse_trace_server_batched.py
@@ -82,16 +82,29 @@ from weave.trace_server.calls_query_builder.calls_query_builder import (
 from weave.trace_server.calls_query_builder.usage_query_builder import (
     build_usage_query,
 )
+from weave.trace_server.clickhouse.schema_converters import (
+    ch_call_dict_to_call_schema_dict,
+    ch_call_to_row,
+    ch_complete_call_to_row,
+    ch_obj_to_obj_schema,
+    ch_table_stats_to_table_stats_schema,
+    complete_call_to_ch_insertable,
+    end_call_for_insert_to_ch_insertable,
+    find_call_descendants,
+    get_kind,
+    maybe_enqueue_minimal_call_end,
+    start_call_for_insert_to_ch_insertable,
+    start_call_insertable_to_complete_start,
+    start_end_calls_to_ch_complete_insertable,
+)
 from weave.trace_server.clickhouse.utilities import (
     any_value_to_dump,
     convert_to_insert_too_large,
     datetime_to_microseconds,
-    dict_dump_to_dict,
     dict_value_to_dump,
     ensure_datetimes_have_tz,
     ensure_datetimes_have_tz_strict,
     log_and_raise_insert_error,
-    nullable_any_dump_to_any,
     num_bytes,
     process_parameters,
     should_retry_empty_query,
@@ -240,7 +253,6 @@ from weave.trace_server.trace_server_common import (
     determine_call_status,
     get_nested_key,
     hydrate_calls_with_feedback,
-    make_derived_summary_fields,
     make_feedback_query_req,
     set_nested_key,
 )
@@ -649,7 +661,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         # Build event callbacks (same for both write targets)
         event_callbacks = [
             partial(
-                _maybe_enqueue_minimal_call_end,
+                maybe_enqueue_minimal_call_end,
                 self.kafka_producer,
                 end_call.project_id,
                 end_call.id,
@@ -662,8 +674,8 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         # Convert and insert based on write target
         if write_target == WriteTarget.CALLS_COMPLETE:
             rows = [
-                _ch_complete_call_to_row(
-                    _start_end_calls_to_ch_complete_insertable(start, end)
+                ch_complete_call_to_row(
+                    start_end_calls_to_ch_complete_insertable(start, end)
                 )
                 for start, end in calls
             ]
@@ -672,13 +684,9 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
             rows = []
             for start, end in calls:
                 rows.append(
-                    _ch_call_to_row(
-                        _start_call_for_insert_to_ch_insertable_start_call(start)
-                    )
+                    ch_call_to_row(start_call_for_insert_to_ch_insertable(start))
                 )
-                rows.append(
-                    _ch_call_to_row(_end_call_for_insert_to_ch_insertable_end_call(end))
-                )
+                rows.append(ch_call_to_row(end_call_for_insert_to_ch_insertable(end)))
             self._insert_call_batch(rows)
 
         # Run callbacks and flush
@@ -772,7 +780,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         set_current_span_dd_tags({"weave_trace_server.insert_call_count": 1})
 
         req = process_call_req_to_content(req, self)
-        ch_call = _start_call_for_insert_to_ch_insertable_start_call(req.start)
+        ch_call = start_call_for_insert_to_ch_insertable(req.start)
 
         # Check write target - v1 call_start cannot write to calls_complete
         write_target = self.table_routing_resolver.resolve_v1_write_target(
@@ -801,7 +809,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         # This does validation and conversion of the input data as well
         # as enforcing business rules and defaults
         req = process_call_req_to_content(req, self)
-        ch_call = _end_call_for_insert_to_ch_insertable_end_call(req.end)
+        ch_call = end_call_for_insert_to_ch_insertable(req.end)
 
         # Check write target - v1 call_end cannot write to calls_complete
         write_target = self.table_routing_resolver.resolve_v1_write_target(
@@ -816,7 +824,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         self._insert_call(ch_call)
 
         if publish:
-            _maybe_enqueue_minimal_call_end(
+            maybe_enqueue_minimal_call_end(
                 self.kafka_producer,
                 req.end.project_id,
                 req.end.id,
@@ -865,13 +873,13 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
                     self.ch_client,
                 )
 
-                ch_call = _complete_call_to_ch_insertable(processed_complete_call)
+                ch_call = complete_call_to_ch_insertable(processed_complete_call)
                 if write_target == WriteTarget.CALLS_COMPLETE:
                     self._insert_call_complete(ch_call)
                 else:
                     self._insert_call_to_v1(ch_call)
 
-                _maybe_enqueue_minimal_call_end(
+                maybe_enqueue_minimal_call_end(
                     self.kafka_producer,
                     processed_complete_call.project_id,
                     processed_complete_call.id,
@@ -887,14 +895,14 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         their start to be visible immediately in the UI.
         """
         start_req = process_call_req_to_content(tsi.CallStartReq(start=req.start), self)
-        ch_start = _start_call_for_insert_to_ch_insertable_start_call(start_req.start)
+        ch_start = start_call_for_insert_to_ch_insertable(start_req.start)
 
         write_target = self.table_routing_resolver.resolve_v2_write_target(
             ch_start.project_id,
             self.ch_client,
         )
         if write_target == WriteTarget.CALLS_COMPLETE:
-            ch_complete_start = _start_call_insertable_to_complete_start(ch_start)
+            ch_complete_start = start_call_insertable_to_complete_start(ch_start)
             self._insert_call_complete(ch_complete_start)
         else:
             self._insert_call(ch_start)
@@ -925,12 +933,12 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         if write_target == WriteTarget.CALLS_COMPLETE:
             self._update_call_end_in_calls_complete(req.end)
         elif write_target == WriteTarget.CALLS_MERGED:
-            ch_end = _end_call_for_insert_to_ch_insertable_end_call(req.end)
+            ch_end = end_call_for_insert_to_ch_insertable(req.end)
             self._insert_call(ch_end)
             if self._flush_immediately:
                 self._flush_calls()
 
-        _maybe_enqueue_minimal_call_end(
+        maybe_enqueue_minimal_call_end(
             self.kafka_producer,
             req.end.project_id,
             req.end.id,
@@ -1423,7 +1431,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
             set_current_span_dd_tags({"expand_columns": "true"})
 
         def row_to_call_schema_dict(row: tuple[Any, ...]) -> dict[str, Any]:
-            return _ch_call_dict_to_call_schema_dict(
+            return ch_call_dict_to_call_schema_dict(
                 dict(zip(select_columns, row, strict=False))
             )
 
@@ -1478,7 +1486,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         try:
             for row in raw_res:
                 yield tsi.CallSchema.model_validate(
-                    _ch_call_dict_to_call_schema_dict(
+                    ch_call_dict_to_call_schema_dict(
                         dict(zip(select_columns, row, strict=True))
                     )
                 )
@@ -1843,7 +1851,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
                 deleted_at=obj.deleted_at,
             )
 
-        obj_schema = _ch_obj_to_obj_schema(obj)
+        obj_schema = ch_obj_to_obj_schema(obj)
         if req.include_tags_and_aliases:
             set_root_span_dd_tags({"include_tags_and_aliases": True})
             self._enrich_objs_with_tags_and_aliases(req.project_id, [obj_schema])
@@ -1890,7 +1898,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         object_query_builder.set_include_deleted(include_deleted=False)
         object_query_builder.include_storage_size = req.include_storage_size or False
         objs = self._select_objs_query(object_query_builder, metadata_only)
-        obj_schemas = [_ch_obj_to_obj_schema(obj) for obj in objs]
+        obj_schemas = [ch_obj_to_obj_schema(obj) for obj in objs]
         if req.include_tags_and_aliases:
             set_root_span_dd_tags({"include_tags_and_aliases": True})
             self._enrich_objs_with_tags_and_aliases(req.project_id, obj_schemas)
@@ -2510,7 +2518,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         query_result = self.ch_client.query(query, parameters=parameters)
 
         tables = [
-            _ch_table_stats_to_table_stats_schema(row)
+            ch_table_stats_to_table_stats_schema(row)
             for row in query_result.result_rows
         ]
 
@@ -6016,7 +6024,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
                 exception=exception,
                 wb_user_id=req.wb_user_id,
             )
-            ch_call = _complete_call_to_ch_insertable(completed)
+            ch_call = complete_call_to_ch_insertable(completed)
             self._insert_call_complete(ch_call)
         else:
             # Write to call_parts/calls_merged via start/end pattern
@@ -6035,7 +6043,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
                 },
                 attributes={},
             )
-            start_call = _start_call_for_insert_to_ch_insertable_start_call(start)
+            start_call = start_call_for_insert_to_ch_insertable(start)
             end = tsi.EndedCallSchemaForInsert(
                 project_id=req.project_id,
                 id=start_call.id,
@@ -6045,7 +6053,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
             )
             if exception:
                 end.exception = exception
-            end_call = _end_call_for_insert_to_ch_insertable_end_call(end)
+            end_call = end_call_for_insert_to_ch_insertable(end)
             calls: list[CallStartCHInsertable | CallEndCHInsertable] = [
                 start_call,
                 end_call,
@@ -6148,10 +6156,10 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
                 inputs=tracked_inputs,
                 attributes={},
             )
-            start_call = _start_call_for_insert_to_ch_insertable_start_call(start)
+            start_call = start_call_for_insert_to_ch_insertable(start)
             # Insert immediately so that callers can see the call in progress
             if write_target == WriteTarget.CALLS_COMPLETE:
-                ch_complete_start = _start_call_insertable_to_complete_start(start_call)
+                ch_complete_start = start_call_insertable_to_complete_start(start_call)
                 self._insert_call_complete(ch_complete_start)
             else:
                 self._insert_call(start_call)
@@ -6279,7 +6287,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
             inputs=input_data,
             attributes={},
         )
-        start_call = _start_call_for_insert_to_ch_insertable_start_call(start)
+        start_call = start_call_for_insert_to_ch_insertable(start)
 
         end = tsi.EndedCallSchemaForInsert(
             project_id=req.project_id,
@@ -6295,7 +6303,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
         if "error" in res.response:
             end.exception = res.response["error"]
 
-        end_call = _end_call_for_insert_to_ch_insertable_end_call(end)
+        end_call = end_call_for_insert_to_ch_insertable(end)
         calls: list[CallStartCHInsertable | CallEndCHInsertable] = [
             start_call,
             end_call,
@@ -6782,7 +6790,7 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
             return
 
         # Safety net: enforce sentinel conversion on every row before hitting ClickHouse.
-        # Callers (e.g. _insert_call_complete, _ch_complete_call_to_row) should have
+        # Callers (e.g. _insert_call_complete, ch_complete_call_to_row) should have
         # already applied to_ch_value(), but a single missed call-site produces the
         # cryptic "Invalid None value in non-Nullable column" error.  Applying here as
         # well is cheap (idempotent for non-None values) and makes new code paths safe
@@ -6866,434 +6874,6 @@ class ClickHouseTraceServer(tsi.FullTraceServerInterface):
             }
         )
         return final_batch
-
-
-def _ch_call_dict_to_call_schema_dict(ch_call_dict: dict) -> dict:
-    summary = nullable_any_dump_to_any(ch_call_dict.get("summary_dump"))
-    started_at = ensure_datetimes_have_tz(ch_call_dict.get("started_at"))
-
-    # Convert sentinel values back to None for all sentinel-tracked fields.
-    # This handles both the calls_merged Nullable path (returns None) and the
-    # calls_complete non-nullable path (returns sentinel -> converted to None).
-    sv: dict[str, Any] = {}
-    for field in ch_sentinel_values.ALL_SENTINEL_FIELDS:
-        raw = ch_call_dict.get(field)
-        val = ch_sentinel_values.from_ch_value(field, raw)
-        if field in ch_sentinel_values.SENTINEL_DATETIME_FIELDS:
-            val = ensure_datetimes_have_tz(val)
-        sv[field] = val
-
-    ended_at = sv["ended_at"]
-    display_name = sv["display_name"]
-    exception = sv["exception"]
-    otel_dump = sv["otel_dump"]
-
-    # Load attributes from attributes_dump
-    attributes = dict_dump_to_dict(ch_call_dict.get("attributes_dump", "{}"))
-
-    # For backwards/future compatibility: inject otel_dump into attributes if present
-    # Legacy trace servers stored all otel info in attributes, clients expect it
-    # TODO(gst): consider returning the raw otel column and reconstructing client side
-    if otel_dump:
-        attributes["otel_span"] = dict_dump_to_dict(otel_dump)
-
-    return {
-        "project_id": ch_call_dict.get("project_id"),
-        "id": ch_call_dict.get("id"),
-        "trace_id": ch_call_dict.get("trace_id"),
-        "parent_id": sv["parent_id"],
-        "thread_id": sv["thread_id"],
-        "turn_id": sv["turn_id"],
-        "op_name": ch_call_dict.get("op_name"),
-        "started_at": started_at,
-        "ended_at": ended_at,
-        "attributes": attributes,
-        "inputs": dict_dump_to_dict(ch_call_dict.get("inputs_dump", "{}")),
-        "output": nullable_any_dump_to_any(ch_call_dict.get("output_dump")),
-        "summary": make_derived_summary_fields(
-            summary=summary or {},
-            op_name=ch_call_dict.get("op_name", ""),
-            started_at=started_at,
-            ended_at=ended_at,
-            exception=exception,
-            display_name=display_name,
-        ),
-        "exception": exception,
-        "wb_run_id": sv["wb_run_id"],
-        "wb_run_step": ch_call_dict.get("wb_run_step"),
-        "wb_run_step_end": ch_call_dict.get("wb_run_step_end"),
-        "wb_user_id": sv["wb_user_id"],
-        "display_name": display_name,
-        "storage_size_bytes": ch_call_dict.get("storage_size_bytes"),
-        "total_storage_size_bytes": ch_call_dict.get("total_storage_size_bytes"),
-    }
-
-
-def _ch_obj_to_obj_schema(ch_obj: SelectableCHObjSchema) -> tsi.ObjSchema:
-    return tsi.ObjSchema(
-        project_id=ch_obj.project_id,
-        object_id=ch_obj.object_id,
-        created_at=ensure_datetimes_have_tz(ch_obj.created_at),
-        wb_user_id=ch_obj.wb_user_id,
-        version_index=ch_obj.version_index,
-        is_latest=ch_obj.is_latest,
-        digest=ch_obj.digest,
-        kind=ch_obj.kind,
-        base_object_class=ch_obj.base_object_class,
-        leaf_object_class=ch_obj.leaf_object_class,
-        val=json.loads(ch_obj.val_dump),
-        size_bytes=ch_obj.size_bytes,
-    )
-
-
-def _ch_table_stats_to_table_stats_schema(
-    ch_table_stats_row: Sequence[Any],
-) -> tsi.TableStatsRow:
-    # Unpack the row with a default for the third value if it doesn't exist
-    row_tuple = tuple(ch_table_stats_row)
-    digest, count = row_tuple[:2]
-    storage_size_bytes = row_tuple[2] if len(row_tuple) > 2 else cast(Any, None)
-
-    return tsi.TableStatsRow(
-        count=count,
-        digest=digest,
-        storage_size_bytes=storage_size_bytes,
-    )
-
-
-def _ch_call_to_row(ch_call: CallCHInsertable) -> list[Any]:
-    """Convert a CH insertable call to a row for batch insertion with the correct defaults."""
-    call_dict = ch_call.model_dump()
-    return [call_dict.get(col) for col in ALL_CALL_INSERT_COLUMNS]
-
-
-def _start_call_for_insert_to_ch_insertable_start_call(
-    start_call: tsi.StartedCallSchemaForInsert,
-) -> CallStartCHInsertable:
-    # Note: it is technically possible for the user to mess up and provide the
-    # wrong trace id (one that does not match the parent_id)!
-    call_id = start_call.id or generate_id()
-    trace_id = start_call.trace_id or generate_id()
-    # Process inputs for base64 content if trace_server is provided
-    inputs = start_call.inputs
-    input_refs = extract_refs_from_values(inputs)
-
-    otel_dump_str = None
-    if start_call.otel_dump is not None:
-        otel_dump_str = dict_value_to_dump(start_call.otel_dump)
-
-    return CallStartCHInsertable(
-        project_id=start_call.project_id,
-        id=call_id,
-        trace_id=trace_id,
-        parent_id=start_call.parent_id,
-        thread_id=start_call.thread_id,
-        turn_id=start_call.turn_id,
-        op_name=start_call.op_name,
-        started_at=start_call.started_at,
-        attributes_dump=dict_value_to_dump(start_call.attributes),
-        inputs_dump=dict_value_to_dump(inputs),
-        input_refs=input_refs,
-        otel_dump=otel_dump_str,
-        wb_run_id=start_call.wb_run_id,
-        wb_run_step=start_call.wb_run_step,
-        wb_user_id=start_call.wb_user_id,
-        display_name=start_call.display_name,
-    )
-
-
-def _start_call_insertable_to_complete_start(
-    ch_start: CallStartCHInsertable,
-) -> CallCompleteCHInsertable:
-    """Convert a start-only call into a calls_complete insertable row.
-
-    Args:
-        ch_start: The start-only ClickHouse insertable call.
-
-    Returns:
-        CallCompleteCHInsertable: A calls_complete insertable row with an empty end.
-
-    Examples:
-        >>> import datetime
-        >>> ch_start = CallStartCHInsertable(
-        ...     project_id="entity/project",
-        ...     id="call-id",
-        ...     trace_id="trace-id",
-        ...     op_name="op",
-        ...     started_at=datetime.datetime(2024, 1, 1),
-        ...     attributes_dump="{}",
-        ...     inputs_dump="{}",
-        ...     input_refs=[],
-        ...     output_refs=[],
-        ... )
-        >>> complete = _start_call_insertable_to_complete_start(ch_start)
-        >>> complete.ended_at is None
-        True
-    """
-    return CallCompleteCHInsertable(
-        project_id=ch_start.project_id,
-        id=ch_start.id,
-        trace_id=ch_start.trace_id,
-        parent_id=ch_start.parent_id,
-        thread_id=ch_start.thread_id,
-        turn_id=ch_start.turn_id,
-        op_name=ch_start.op_name,
-        display_name=ch_start.display_name,
-        started_at=ch_start.started_at,
-        ended_at=None,
-        exception=None,
-        attributes_dump=ch_start.attributes_dump,
-        inputs_dump=ch_start.inputs_dump,
-        input_refs=ch_start.input_refs,
-        output_dump=any_value_to_dump(None),
-        summary_dump=dict_value_to_dump({}),
-        otel_dump=ch_start.otel_dump,
-        output_refs=ch_start.output_refs,
-        wb_user_id=ch_start.wb_user_id,
-        wb_run_id=ch_start.wb_run_id,
-        wb_run_step=ch_start.wb_run_step,
-        wb_run_step_end=None,
-    )
-
-
-def _end_call_for_insert_to_ch_insertable_end_call(
-    end_call: tsi.EndedCallSchemaForInsert,
-) -> CallEndCHInsertable:
-    # Note: it is technically possible for the user to mess up and provide the
-    # wrong trace id (one that does not match the parent_id)!
-
-    output = end_call.output
-    output_refs = extract_refs_from_values(output)
-
-    return CallEndCHInsertable(
-        project_id=end_call.project_id,
-        id=end_call.id,
-        exception=end_call.exception,
-        ended_at=end_call.ended_at,
-        summary_dump=dict_value_to_dump(dict(end_call.summary)),
-        output_dump=any_value_to_dump(output),
-        output_refs=output_refs,
-        wb_run_step_end=end_call.wb_run_step_end,
-    )
-
-
-def _start_end_calls_to_ch_complete_insertable(
-    start_call: tsi.StartedCallSchemaForInsert,
-    end_call: tsi.EndedCallSchemaForInsert,
-) -> CallCompleteCHInsertable:
-    """Combine start and end call data into a CallCompleteCHInsertable.
-
-    Used by OTel export when writing to the calls_complete table.
-
-    Args:
-        start_call: The start call data.
-        end_call: The end call data.
-
-    Returns:
-        CallCompleteCHInsertable: A complete call ready for insertion.
-    """
-    call_id = start_call.id or generate_id()
-    trace_id = start_call.trace_id or generate_id()
-
-    inputs = start_call.inputs
-    input_refs = extract_refs_from_values(inputs)
-
-    output = end_call.output
-    output_refs = extract_refs_from_values(output)
-
-    otel_dump_str = None
-    if start_call.otel_dump is not None:
-        otel_dump_str = dict_value_to_dump(start_call.otel_dump)
-
-    return CallCompleteCHInsertable(
-        project_id=start_call.project_id,
-        id=call_id,
-        trace_id=trace_id,
-        parent_id=start_call.parent_id,
-        thread_id=start_call.thread_id,
-        turn_id=start_call.turn_id,
-        op_name=start_call.op_name,
-        display_name=start_call.display_name,
-        started_at=start_call.started_at,
-        ended_at=end_call.ended_at,
-        exception=end_call.exception,
-        attributes_dump=dict_value_to_dump(start_call.attributes),
-        inputs_dump=dict_value_to_dump(inputs),
-        input_refs=input_refs,
-        output_dump=any_value_to_dump(output),
-        summary_dump=dict_value_to_dump(dict(end_call.summary)),
-        otel_dump=otel_dump_str,
-        output_refs=output_refs,
-        wb_user_id=start_call.wb_user_id,
-        wb_run_id=start_call.wb_run_id,
-        wb_run_step=start_call.wb_run_step,
-        wb_run_step_end=end_call.wb_run_step_end,
-    )
-
-
-def _ch_complete_call_to_row(ch_call: CallCompleteCHInsertable) -> list[Any]:
-    """Convert a CallCompleteCHInsertable to a row for batch insertion."""
-    call_dict = ch_call.model_dump()
-    return [
-        ch_sentinel_values.to_ch_value(col, call_dict.get(col))
-        for col in ALL_CALL_COMPLETE_INSERT_COLUMNS
-    ]
-
-
-def _maybe_enqueue_minimal_call_end(
-    kafka_producer: KafkaProducer | None,
-    project_id: str,
-    id: str,
-    ended_at: datetime.datetime,
-    flush_immediately: bool = False,
-) -> None:
-    """Enqueue a minimal call end event to Kafka if online eval is enabled.
-
-    This is used for online eval triggers where we only need the call identity,
-    not the full payload. Large fields (output, summary, exception) are stripped.
-
-    Args:
-        kafka_producer: The Kafka producer to use.
-        project_id: The project ID.
-        id: The call ID.
-        ended_at: The call end timestamp.
-        flush_immediately: Whether to flush the producer immediately.
-    """
-    if kafka_producer is None:
-        return
-
-    minimal_end = tsi.EndedCallSchemaForInsert(
-        project_id=project_id,
-        id=id,
-        ended_at=ended_at,
-        output=None,
-        summary={},
-        exception=None,
-    )
-    kafka_producer.produce_call_end(minimal_end, flush_immediately)
-
-
-def _complete_call_to_ch_insertable(
-    complete_call: tsi.CompletedCallSchemaForInsert,
-) -> CallCompleteCHInsertable:
-    """Convert a completed call schema to a ClickHouse insertable format.
-
-    Args:
-        complete_call: The completed call schema from the API.
-
-    Returns:
-        CallCompleteCHInsertable: The ClickHouse insertable representation.
-    """
-    inputs = complete_call.inputs
-    input_refs = extract_refs_from_values(inputs)
-
-    output = complete_call.output
-    output_refs = extract_refs_from_values(output)
-
-    otel_dump_str = None
-    if complete_call.otel_dump is not None:
-        otel_dump_str = dict_value_to_dump(complete_call.otel_dump)
-
-    return CallCompleteCHInsertable(
-        project_id=complete_call.project_id,
-        id=complete_call.id,
-        trace_id=complete_call.trace_id,
-        parent_id=complete_call.parent_id,
-        thread_id=complete_call.thread_id,
-        turn_id=complete_call.turn_id,
-        op_name=complete_call.op_name,
-        display_name=complete_call.display_name,
-        started_at=complete_call.started_at,
-        ended_at=complete_call.ended_at,
-        exception=complete_call.exception,
-        attributes_dump=dict_value_to_dump(complete_call.attributes),
-        inputs_dump=dict_value_to_dump(inputs),
-        input_refs=input_refs,
-        output_dump=any_value_to_dump(output),
-        summary_dump=dict_value_to_dump(dict(complete_call.summary)),
-        otel_dump=otel_dump_str,
-        output_refs=output_refs,
-        wb_user_id=complete_call.wb_user_id,
-        wb_run_id=complete_call.wb_run_id,
-        wb_run_step=complete_call.wb_run_step,
-        wb_run_step_end=complete_call.wb_run_step_end,
-    )
-
-
-# def _partial_obj_schema_to_ch_obj(
-#     partial_obj: tsi.ObjSchemaForInsert,
-# ) -> ObjCHInsertable:
-#     version_hash = version_hash_for_object(partial_obj)
-
-#     return ObjCHInsertable(
-#         id=uuid.uuid4(),
-#         project_id=partial_obj.project_id,
-#         name=partial_obj.name,
-#         type="unknown",
-#         refs=[],
-#         val=json.dumps(partial_obj.val),
-#     )
-
-
-def get_type(val: Any) -> str:
-    if val is None:
-        return "none"
-    elif isinstance(val, dict):
-        if "_type" in val:
-            if "weave_type" in val:
-                return val["weave_type"]["type"]
-            return val["_type"]
-        return "dict"
-    elif isinstance(val, list):
-        return "list"
-    return "unknown"
-
-
-def get_kind(val: Any) -> str:
-    val_type = get_type(val)
-    if val_type == "Op":
-        return "op"
-    return "object"
-
-
-@ddtrace.tracer.wrap(name="clickhouse_trace_server_batched.find_call_descendants")
-def find_call_descendants(
-    root_ids: list[str],
-    all_calls: list[tsi.CallSchema],
-) -> list[str]:
-    set_current_span_dd_tags(
-        {
-            "clickhouse_trace_server_batched.find_call_descendants.root_ids_count": str(
-                len(root_ids)
-            ),
-            "clickhouse_trace_server_batched.find_call_descendants.all_calls_count": str(
-                len(all_calls)
-            ),
-        }
-    )
-    # make a map of call_id to children list
-    children_map = defaultdict(list)
-    for call in all_calls:
-        if call.parent_id is not None:
-            children_map[call.parent_id].append(call.id)
-
-    # do DFS to get all descendants
-    def find_all_descendants(root_ids: list[str]) -> set[str]:
-        descendants = set()
-        stack = root_ids
-
-        while stack:
-            current_id = stack.pop()
-            if current_id not in descendants:
-                descendants.add(current_id)
-                stack += children_map.get(current_id, [])
-
-        return descendants
-
-    # Find descendants for each initial id
-    descendants = find_all_descendants(root_ids)
-
-    return list(descendants)
 
 
 def _update_metadata_from_chunk(
@@ -7462,7 +7042,7 @@ def _create_tracked_stream_wrapper(
             if end_call_handler is not None:
                 end_call_handler(end)
             else:
-                end_call_ch = _end_call_for_insert_to_ch_insertable_end_call(end)
+                end_call_ch = end_call_for_insert_to_ch_insertable(end)
                 insert_call(end_call_ch)
 
     return _stream_wrapper()
